Remove enricher and indexer from setup, preparation to hosting  it in a separate repository.
