+++ addons/input_module/localfiles/enricher/enricher.py
@@ -1,26 +1,22 @@
 #!/usr/bin/env python3
-
-import logging
-import time
-import threading
-import queue
+import os
+import argparse
+import asyncio
 import enum
+import json
+import logging
+import socket
+import sys
+import tempfile
+from typing import Dict, Optional, Any
 
+from musicbrainz_plugin import MusicBrainzPlugin
+from acoustid_plugin import AcoustIdPlugin
+from wikidata_plugin import WikidataPlugin
+from deezer_plugin import DeezerPlugin
+from enricher_db import AsyncEnricherDb
 
-try:
-    from musicbrainz_plugin import MusicBrainzPlugin
-    from acoustid_plugin import AcoustIdPlugin
-    from wikidata_plugin import WikidataPlugin
-    from deezer_plugin import DeezerPlugin
-    from enricher_db import EnricherDb
-except ImportError:
-    from .musicbrainz_plugin import MusicBrainzPlugin
-    from .acoustid_plugin import AcoustIdPlugin
-    from .wikidata_plugin import WikidataPlugin
-    from .deezer_plugin import DeezerPlugin
-    from .enricher_db import EnricherDb
-
-logger = logging.getLogger(__name__.split(".")[-1])
+logger = logging.getLogger("enricher")
 
 # Set MusicBrainzNGS log level to warning to reduce verbosity
 musicbrainz_logger = logging.getLogger("musicbrainzngs")
@@ -39,10 +35,13 @@ TRACK_REQUIRED_FIELDS = [
     "track_number",
 ]
 
+# Socket path for IPC
+SOCKET_PATH = os.path.join(tempfile.gettempdir(), "kalinka-enricher.sock")
+
 # Global variables to manage enricher state
-_enricher_thread = None
-_enricher_instance = None
-_enricher_queue = queue.Queue()
+_server = None
+_enricher_task: Optional[asyncio.Task] = None
+_enricher_queue: asyncio.Queue = asyncio.Queue()
 
 
 class EnrichmentStatus(enum.IntEnum):
@@ -54,30 +53,30 @@ class EnrichmentStatus(enum.IntEnum):
 
 
 class MetadataEnricher:
-    """Main enricher class that processes database entries"""
+    """Main enricher class that processes database entries using async"""
 
-    def __init__(self, config, db_manager):
+    def __init__(self, config, db_manager: AsyncEnricherDb):
         self.config = config
         self.db_manager = db_manager
         self.running = False
-        self.lock = threading.Lock()
+        self.lock = asyncio.Lock()
         self.plugins = []
 
         # The order of plugins matters for the enrichment process
         # as the first one found a match will be used
-        if config["enricher.plugins.acoustid.enabled"]:
+        if config.get("enricher.plugins.acoustid.enabled", False):
             self.plugins.append(AcoustIdPlugin(config, self.db_manager))
 
-        if config["enricher.plugins.musicbrainz.enabled"]:
+        if config.get("enricher.plugins.musicbrainz.enabled", False):
             self.plugins.append(MusicBrainzPlugin(config, self.db_manager))
 
-        if config["enricher.plugins.wikidata.enabled"]:
+        if config.get("enricher.plugins.wikidata.enabled", False):
             self.plugins.append(WikidataPlugin(config, self.db_manager))
 
-        if config["enricher.plugins.deezer.enabled"]:
+        if config.get("enricher.plugins.deezer.enabled", False):
             self.plugins.append(DeezerPlugin(config, self.db_manager))
 
-    def process_changed_items(self, changed_items):
+    async def process_changed_items(self, changed_items):
         """Process specific items that were changed by the indexer"""
         if not changed_items:
             return
@@ -90,62 +89,63 @@ class MetadataEnricher:
         # Process specific artists
         for artist_id in changed_items.get("artists", []):
             if artist_id and artist_id != "unknown_artist":
-                artist = self.db_manager.get_artist_by_id(artist_id)
+                artist = await self.db_manager.get_artist_by_id(artist_id)
                 if artist:
-                    self._enrich_artist(artist)
+                    await self._enrich_artist(artist)
 
         # Process specific albums
         for album_id in changed_items.get("albums", []):
             if album_id and album_id != "unknown_album":
-                album = self.db_manager.get_album_by_id(album_id)
+                album = await self.db_manager.get_album_by_id(album_id)
                 if album:
-                    self._enrich_album(album)
+                    await self._enrich_album(album)
 
         # Process specific tracks
         for track_id in changed_items.get("tracks", []):
             if track_id:
-                track = self.db_manager.get_track_by_id(track_id)
+                track = await self.db_manager.get_track_by_id(track_id)
                 if track:
-                    self._enrich_track(track)
+                    await self._enrich_track(track)
 
-    def start(self):
+    async def start(self):
         """Start the enricher process for general enrichment"""
-        if self.running:
-            logger.warning("Enricher already running, skipping")
-            return
-
-        with self.lock:
+        async with self.lock:
+            if self.running:
+                logger.warning("Enricher already running, skipping")
+                return
             self.running = True
-            try:
-                self.run_enrichment()
-                logger.info("Enricher process completed")
-            except Exception as e:
-                logger.error(f"Error running enricher: {str(e)}")
-            finally:
+
+        try:
+            await self.run_enrichment()
+            logger.info("Enricher process completed")
+        except Exception as e:
+            logger.exception(f"Error running enricher: {e}")
+        finally:
+            async with self.lock:
                 self.running = False
 
-    def run_enrichment(self):
+    async def run_enrichment(self):
         """Run the enrichment process"""
         # Process artists
         logger.info("Processing artists for enrichment")
-        self._process_artists()
+        await self._process_artists()
 
         # Process albums
         logger.info("Processing albums for enrichment")
-        self._process_albums()
+        await self._process_albums()
 
         # Process tracks
         logger.info("Processing tracks for enrichment")
-        self._process_tracks()
+        await self._process_tracks()
 
-    def _process_artists(self, batch_size=10):
+    async def _process_artists(self, batch_size=10):
         """Process non-enriched artists"""
-        artists = self.db_manager.get_non_enriched_artists(batch_size)
+        artists = await self.db_manager.get_non_enriched_artists(batch_size)
 
         for artist in artists:
-            self._enrich_artist(artist)
+            await self._enrich_artist(artist)
 
-    def _enrich_artist(self, artist):
+    async def _enrich_artist(self, artist):
         """Enrich a single artist"""
         updated_artist = artist.copy()  # Make a copy to carry updates between plugins
         had_updates = False
@@ -157,7 +157,7 @@ class MetadataEnricher:
         if is_fully_enriched:
             logger.debug(f"Artist {artist['id']} already has all required fields")
             updated_artist["enriched"] = EnrichmentStatus.ENRICHED
-            self.db_manager.update_artist(
+            await self.db_manager.update_artist(
                 artist["id"], {"enriched": EnrichmentStatus.ENRICHED}
             )
             return
@@ -166,7 +166,7 @@ class MetadataEnricher:
             if not plugin.can_enrich_artist():
                 continue
 
-            result = plugin.enrich_artist(updated_artist)
+            result = await plugin.enrich_artist(updated_artist)
             if result and "updates" in result:
                 # Apply updates to our working copy
                 updated_artist.update(result["updates"])
@@ -195,16 +195,16 @@ class MetadataEnricher:
 
         # Only update the database once at the end if we had any updates
         if had_updates:
-            self.db_manager.update_artist(artist["id"], updated_artist)
+            await self.db_manager.update_artist(artist["id"], updated_artist)
 
-    def _process_albums(self, batch_size=500):
+    async def _process_albums(self, batch_size=500):
         """Process non-enriched albums"""
-        albums = self.db_manager.get_non_enriched_albums(batch_size)
+        albums = await self.db_manager.get_non_enriched_albums(batch_size)
 
         for album in albums:
-            self._enrich_album(album)
+            await self._enrich_album(album)
 
-    def _enrich_album(self, album):
+    async def _enrich_album(self, album):
         """Enrich a single album"""
         updated_album = album.copy()  # Make a copy to carry updates between plugins
         had_updates = False
@@ -216,7 +216,7 @@ class MetadataEnricher:
         if is_fully_enriched:
             logger.debug(f"Album {album['id']} already has all required fields")
             updated_album["enriched"] = EnrichmentStatus.ENRICHED
-            self.db_manager.update_album(
+            await self.db_manager.update_album(
                 album["id"], {"enriched": EnrichmentStatus.ENRICHED}
             )
             return
@@ -228,7 +228,7 @@ class MetadataEnricher:
             logger.debug(
                 f"Enriching album {album['id']} with {plugin.__class__.__name__}"
             )
-            result = plugin.enrich_album(updated_album)
+            result = await plugin.enrich_album(updated_album)
             logger.debug(f"Result from {plugin.__class__.__name__}: {result}")
             if result and "updates" in result:
                 logger.debug(f"Updates found: {result['updates']}")
@@ -259,18 +259,17 @@ class MetadataEnricher:
 
         # Only update the database once at the end if we had any updates
         if had_updates:
-            self.db_manager.update_album(album["id"], updated_album)
+            await self.db_manager.update_album(album["id"], updated_album)
 
-    def _process_tracks(self, batch_size=500):
+    async def _process_tracks(self, batch_size=500):
         """Process non-enriched tracks"""
-        tracks = self.db_manager.get_non_enriched_tracks(batch_size)
+        tracks = await self.db_manager.get_non_enriched_tracks(batch_size)
 
         for track in tracks:
-            self._enrich_track(track)
+            await self._enrich_track(track)
 
-    def _enrich_track(self, track):
+    async def _enrich_track(self, track):
         """Enrich a single track"""
-
         updated_track = track.copy()  # Make a copy to carry updates between plugins
         had_updates = False
 
@@ -281,7 +280,7 @@ class MetadataEnricher:
         if is_fully_enriched:
             logger.debug(f"Track {track['id']} already has all required fields")
             updated_track["enriched"] = EnrichmentStatus.ENRICHED
-            self.db_manager.update_track(
+            await self.db_manager.update_track(
                 track["id"], {"enriched": EnrichmentStatus.ENRICHED}
             )
             return
@@ -300,7 +299,7 @@ class MetadataEnricher:
                 )
                 break
 
-            result = plugin.enrich_track(updated_track)
+            result = await plugin.enrich_track(updated_track)
             if result:
                 if "updates" in result:
                     # Apply updates to our working copy
@@ -338,7 +337,10 @@ class MetadataEnricher:
 
         # Only update the database once at the end if we had any updates
         if had_updates:
-            self.db_manager.update_track(track["id"], updated_track)
+            logger.info(
+                f"Updating track {track['name']} with new metadata: {updated_track.keys()}"
+            )
+            await self.db_manager.update_track(track["id"], updated_track)
 
         # If we have entities that need further enrichment, add them to the enricher queue
         if any(entities_for_enrichment.values()):
@@ -357,174 +359,235 @@ class MetadataEnricher:
 
             # Add to the enricher queue
             global _enricher_queue
-            _enricher_queue.put({"changed_items": changed_items})
+            await _enricher_queue.put({"changed_items": changed_items})
 
 
-def _enricher_worker(config, db_manager):
-    """Background worker thread for the enricher"""
-    global _enricher_instance
+async def _enricher_worker(config, db_manager: AsyncEnricherDb):
+    """Background worker task for the enricher"""
 
-    # Create EnricherDb instance
-    enricher_db = EnricherDb(config)
-    _enricher_instance = MetadataEnricher(config, enricher_db)
+    enricher_instance = MetadataEnricher(config, db_manager)
+    enricher_tasks = set()
 
-    logger.info("Metadata enricher thread running")
-    logger.info("Waiting for indexer to complete initial scan")
+    def run_enricher_task(coro):
+        """Run the enricher task"""
+        task = asyncio.create_task(coro)
+        enricher_tasks.add(task)
+        task.add_done_callback(enricher_tasks.discard)
+        return task
 
-    initial_scan_complete = False
+    logger.info("Starting initial enrichment process")
+    run_enricher_task(enricher_instance.start()).add_done_callback(
+        lambda task: logger.info("Initial enrichment process completed")
+    )
+
     # Process queue commands
     while True:
         try:
             # Check for commands
             try:
-                command = _enricher_queue.get(timeout=60)  # Check every minute
-
-                if initial_scan_complete is False:
-                    logger.info("Starting enricher after initial scan")
-                    initial_scan_complete = True
-                    # Start the enricher process
-                    _enricher_instance.start()
-                    logger.info(
-                        "Initial enrichment finished, listening for indexer updates"
-                    )
-                    command = None
+                command = await _enricher_queue.get()
 
                 if command == "stop":
-                    logger.info("Stopping enricher thread")
+                    logger.info("Stopping enricher task")
                     break
                 elif command == "enrich":
                     logger.info("Manual enrichment triggered")
-                    _enricher_instance.start()
+                    run_enricher_task(enricher_instance.start())
                 elif isinstance(command, dict) and "changed_items" in command:
                     # This is a notification from the indexer
-                    logger.info("Processing changed items from indexer")
-                    _enricher_instance.process_changed_items(command["changed_items"])
+                    run_enricher_task(
+                        enricher_instance.process_changed_items(
+                            command["changed_items"]
+                        )
+                    )
 
                 _enricher_queue.task_done()
-            except queue.Empty:
+            except asyncio.TimeoutError:
                 # No command received, just continue
                 pass
 
-            # Sleep to avoid busy-waiting
-            time.sleep(1)
+            # Small delay to avoid busy waiting
+            await asyncio.sleep(0.1)
+        except asyncio.CancelledError:
+            logger.info("Enricher worker cancelled.")
+            break
         except Exception as e:
-            logger.error(f"Error in enricher worker: {str(e)}")
-            time.sleep(5)  # Sleep longer on errors
+            logger.exception(f"Error in enricher worker: {str(e)}")
+            await asyncio.sleep(5)  # Sleep longer on errors
+
+    running_tasks = list(enricher_tasks)
+    for task in running_tasks:
+        if not task.done():
+            task.cancel()
+            try:
+                await task
+            except asyncio.CancelledError:
+                pass
 
 
-def indexer_callback(event):
-    """Callback function to be registered with the indexer"""
-    if _enricher_thread and _enricher_thread.is_alive():
-        logger.info("Received change notification from indexer")
-        _enricher_queue.put(event)
+async def handle_client(reader, writer):
+    """Handle Unix socket client connection"""
+    data = await reader.readline()
+    message = data.decode().strip()
+
+    try:
+        # Try to parse as JSON
+        command = json.loads(message)
+        await _enricher_queue.put(command)
+    except json.JSONDecodeError:
+        # Simple string command
+        await _enricher_queue.put(message)
+
+    writer.close()
+    await writer.wait_closed()
+
+
+def ensure_single_instance():
+    """Try to connect to the existing socket to see if it's already running."""
+    if os.path.exists(SOCKET_PATH):
+        try:
+            # Try to connect to see if the socket is active
+            test_socket = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+            test_socket.connect(SOCKET_PATH)
+            test_socket.close()
+            print("Another instance is already running.")
+            sys.exit(1)
+        except (ConnectionRefusedError, FileNotFoundError):
+            print("Stale socket found. Cleaning up.")
+            os.remove(SOCKET_PATH)  # stale socket
+
+
+async def start_server():
+    """Start the Unix socket server process"""
+    global _server
+
+    if _server and not _server.is_serving():
+        logger.warning("Server already running, not starting another")
+        return _server
+
+    if os.path.exists(SOCKET_PATH):
+        os.remove(SOCKET_PATH)
+
+    _server = await asyncio.start_unix_server(handle_client, path=SOCKET_PATH)
+    logger.info(f"Server started at {SOCKET_PATH}")
+    return _server
+
+
+def start_enricher(config, db_manager: AsyncEnricherDb) -> Optional[asyncio.Task]:
+    """Start the enricher process in a background task"""
+    global _enricher_task
+
+    def log_task_result(task):
+        try:
+            task.result()  # Will raise if task failed
+        except Exception as e:
+            logger.exception(f"Enricher error: {e}")
+
+    if _enricher_task and not _enricher_task.done():
+        logger.warning("Enricher task already running, not starting another")
+        return _enricher_task
+
+    _enricher_task = asyncio.create_task(_enricher_worker(config, db_manager))
+    _enricher_task.add_done_callback(log_task_result)
+
+    logger.info("Started metadata enricher background task")
+    return _enricher_task
+
+
+async def stop_enricher() -> bool:
+    """Stop the enricher task"""
+    global _enricher_task
+    if _enricher_task and not _enricher_task.done():
+        logger.info("Sending stop command to enricher task")
+        try:
+            await _enricher_queue.put("stop")
+            await asyncio.wait_for(_enricher_task, timeout=10.0)
+            return True
+        except asyncio.TimeoutError:
+            logger.warning("Enricher task did not stop in time, cancelling...")
+            _enricher_task.cancel()
+            try:
+                await asyncio.wait_for(_enricher_task, timeout=5.0)
+            except (asyncio.TimeoutError, asyncio.CancelledError):
+                pass
+            return True
+        except Exception as e:
+            logger.exception(f"Error stopping enricher task: {str(e)}")
+            return False
+    elif _enricher_task and _enricher_task.done():
+        logger.info("Enricher task was already done.")
         return True
-    else:
-        logger.warning("Cannot process indexer changes - enricher thread not running")
-        return False
-
-
-def trigger_enrichment():
-    """Manually trigger enrichment (can be called from other modules)"""
-    if _enricher_thread and _enricher_thread.is_alive():
-        logger.info("Triggering manual enrichment")
-        _enricher_queue.put("enrich")
-        return True
-    else:
-        logger.warning("Cannot trigger enrichment - enricher thread not running")
-        return False
-
-
-def start_enricher(config, db_manager):
-    """Start the enricher process in a background thread"""
-    global _enricher_thread
-
-    # Ensure we don't start multiple enricher threads
-    if _enricher_thread and _enricher_thread.is_alive():
-        logger.warning("Enricher thread already running, not starting another")
-        return _enricher_thread
-
-    # Import indexer and register our callback
-    from addons.input_module.localfiles.indexer import register_enricher_callback
-
-    # Start the worker thread
-    _enricher_thread = threading.Thread(
-        target=_enricher_worker, args=(config, db_manager), daemon=True
-    )
-    _enricher_thread.name = "LocalFiles-Enricher"
-    _enricher_thread.start()
-
-    # Register callback with indexer
-    register_enricher_callback(indexer_callback)
-
-    logger.info("Started metadata enricher background thread")
-    return _enricher_thread
-
-
-def stop_enricher():
-    """Stop the enricher thread"""
-    if _enricher_thread and _enricher_thread.is_alive():
-        logger.info("Sending stop command to enricher thread")
-        _enricher_queue.put("stop")
-        _enricher_thread.join(5)
-        return True
+    logger.info("No active enricher task to stop.")
     return False
 
 
-def print_default_config():
-    """Print a default configuration template as JSON"""
-    default_config = {
-        "enricher.plugins.acoustid.enabled": True,
-        "enricher.plugins.acoustid.api_key": "YOUR_ACOUSTID_API_KEY",
-        "enricher.plugins.musicbrainz.enabled": True,
-        "enricher.plugins.musicbrainz.app_name": "YourAppName",
-        "enricher.plugins.musicbrainz.app_version": "1.0",
-        "enricher.plugins.musicbrainz.contact_info": "your@email.com",
-        "enricher.plugins.wikidata.enabled": True,
-        "enricher.plugins.deezer.enabled": True,
-        "database.path": "/path/to/your/database.sqlite",
-    }
-    return json.dumps(default_config, indent=2)
+async def main(config_data: Dict[str, Any]):
+    """Runs the main application logic asynchronously."""
+    global _enricher_task
+
+    ensure_single_instance()
+
+    db_manager = AsyncEnricherDb(config_data)
+    try:
+        await db_manager.init_db()
+        logger.info("Database initialized successfully.")
+    except Exception as e:
+        logger.exception(f"Fatal: Error initializing database: {str(e)}")
+        sys.exit(1)
+
+    shutdown_event = asyncio.Event()
+    loop = asyncio.get_running_loop()
+
+    for sig in (signal.SIGINT, signal.SIGTERM):
+        loop.add_signal_handler(sig, shutdown_event.set)
+
+    _enricher_task = start_enricher(config_data, db_manager)
+    _server = await start_server()
+
+    try:
+        await shutdown_event.wait()
+    except asyncio.CancelledError:
+        logger.info("Server cancelled, shutting down...")
+    except Exception as e:
+        logger.exception(f"Error in main loop: {str(e)}")
+    finally:
+        logger.info("Main async runner initiating shutdown of tasks...")
+
+        if _enricher_task and not _enricher_task.done():
+            await stop_enricher()
+        else:
+            logger.info("Enricher task already finished.")
+
+        logger.info("All background tasks processed for shutdown.")
+
+        if _server:
+            _server.close()
+            await _server.wait_closed()
+
+        if os.path.exists(SOCKET_PATH):
+            os.remove(SOCKET_PATH)
+
+        logger.info("Server shutdown complete.")
 
 
 if __name__ == "__main__":
-    """Run the enricher as a standalone script
+    import signal
 
-    Example usage:
-    # Show default configuration:
-    python -m addons.input_module.localfiles.enricher.enricher --show-default-config
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+    )
 
-    # Run full enrichment process:
-    python -m addons.input_module.localfiles.enricher.enricher --config '{"enricher.plugins.acoustid.enabled": true, "enricher.plugins.musicbrainz.enabled": true, "enricher.plugins.wikidata.enabled": true, "enricher.plugins.deezer.enabled": true, "database.path": "/path/to/db.sqlite"}' --log-level INFO
-
-    # Enrich specific items:
-    python -m addons.input_module.localfiles.enricher.enricher --config '{"enricher.plugins.acoustid.enabled": true, "database.path": "/path/to/db.sqlite"}' --enrich-artist "artist_id_123" --log-level DEBUG
-    """
-
-    import json
-    import sys
-    import argparse
-
-    parser = argparse.ArgumentParser(description="Standalone metadata enricher")
+    parser = argparse.ArgumentParser(description="Async Metadata enricher daemon")
     parser.add_argument(
-        "--config",
-        type=str,
-        help="JSON string containing configuration",
+        "-c", "--config", help="JSON configuration string", required=True
     )
     parser.add_argument(
         "--log-level",
         type=str,
         choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"],
         default="INFO",
-        help="Set logging level",
-    )
-    parser.add_argument(
-        "--enrich-artist", type=str, help="Enrich a specific artist by ID", default=None
-    )
-    parser.add_argument(
-        "--enrich-album", type=str, help="Enrich a specific album by ID", default=None
-    )
-    parser.add_argument(
-        "--enrich-track", type=str, help="Enrich a specific track by ID", default=None
+        help="Set log level",
     )
     parser.add_argument(
         "--show-default-config",
@@ -533,66 +596,49 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
 
+    # Configure logging level
+    logging.getLogger().setLevel(getattr(logging, args.log_level))
+
     # Show default config if requested
     if args.show_default_config:
-        print(print_default_config())
+        default_config = {
+            "enricher.plugins.acoustid.enabled": True,
+            "enricher.plugins.acoustid.api_key": "YOUR_ACOUSTID_API_KEY",
+            "enricher.plugins.musicbrainz.enabled": True,
+            "enricher.plugins.musicbrainz.app_name": "YourAppName",
+            "enricher.plugins.musicbrainz.app_version": "1.0",
+            "enricher.plugins.musicbrainz.contact_info": "your@email.com",
+            "enricher.plugins.wikidata.enabled": True,
+            "enricher.plugins.deezer.enabled": True,
+            "db_path": "/path/to/your/database.sqlite",
+        }
+        print(json.dumps(default_config, indent=2))
         sys.exit(0)
 
-    # Require config if not showing default config
-    if not args.config:
-        parser.error("--config is required unless --show-default-config is used")
-
-    # Configure logging
-    log_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-    logging.basicConfig(level=getattr(logging, args.log_level), format=log_format)
+    logger.info("Loading configuration from command line JSON")
+    try:
+        loaded_config = json.loads(args.config)
+        if (
+            "input_modules" in loaded_config
+            and "localfiles" in loaded_config["input_modules"]
+        ):
+            app_config = loaded_config["input_modules"]["localfiles"]
+            logger.info(f"Loaded config: {app_config}")
+        else:
+            app_config = loaded_config
+            logger.info("Loaded direct config")
+    except json.JSONDecodeError as e:
+        logger.exception(f"Error parsing JSON configuration: {str(e)}")
+        sys.exit(1)
+    except KeyError as e:
+        logger.exception(f"Configuration missing expected keys: {str(e)}")
+        sys.exit(1)
 
     try:
-        # Parse config from JSON string
-        config = json.loads(args.config)
-
-        # Initialize the database manager
-        enricher_db = EnricherDb(config)
-
-        logger.info("Starting enricher in standalone mode")
-
-        # Create the enricher
-        enricher = MetadataEnricher(config, enricher_db)
-
-        # Check if we need to enrich specific items
-        if args.enrich_artist:
-            logger.info(f"Enriching specific artist: {args.enrich_artist}")
-            artist = enricher_db.get_artist_by_id(args.enrich_artist)
-            if artist:
-                enricher._enrich_artist(artist)
-            else:
-                logger.error(f"Artist not found: {args.enrich_artist}")
-
-        elif args.enrich_album:
-            logger.info(f"Enriching specific album: {args.enrich_album}")
-            album = enricher_db.get_album_by_id(args.enrich_album)
-            if album:
-                enricher._enrich_album(album)
-            else:
-                logger.error(f"Album not found: {args.enrich_album}")
-
-        elif args.enrich_track:
-            logger.info(f"Enriching specific track: {args.enrich_track}")
-            track = enricher_db.get_track_by_id(args.enrich_track)
-            if track:
-                enricher._enrich_track(track)
-            else:
-                logger.error(f"Track not found: {args.enrich_track}")
-
-        else:
-            # Run the full enrichment process
-            logger.info("Running full enrichment process")
-            enricher.start()
-
-        logger.info("Enrichment process completed")
-
-    except json.JSONDecodeError:
-        logger.error("Failed to parse config JSON string")
-        sys.exit(1)
+        asyncio.run(main(app_config))
+    except KeyboardInterrupt:
+        logger.info("KeyboardInterrupt received. Exiting.")
     except Exception as e:
-        logger.error(f"Error running enricher: {str(e)}")
-        sys.exit(1)
+        logger.critical(f"Unhandled exception: {e}", exc_info=True)
+    finally:
+        logger.info("Enricher daemon finished.")
