--- addons/input_module/localfiles/enricher/enricher_db_async.py
@@ -1,409 +1,0 @@
-#!/usr/bin/env python3
-import os
-import aiosqlite
-import logging
-import time
-from typing import List, Dict, Optional, Any, Tuple
-
-logger = logging.getLogger(__name__.split(".")[-1])
-
-
-class AsyncEnricherDb:
-    """
-    Asynchronous database manager specifically for the metadata enricher.
-    Handles operations required for enrichment of music metadata.
-    """
-
-    def __init__(self, config):
-        self.db_path = config["db_path"]
-        self.artwork_path = config["artwork_path"]
-
-        # Ensure directories exist
-        os.makedirs(os.path.dirname(self.db_path), exist_ok=True)
-        os.makedirs(self.artwork_path, exist_ok=True)
-        os.makedirs(os.path.join(self.artwork_path, "album"), exist_ok=True)
-        os.makedirs(os.path.join(self.artwork_path, "artist"), exist_ok=True)
-
-    def _get_connection(self):
-        """Get a database connection with row factory"""
-        conn = aiosqlite.connect(self.db_path)
-
-        return conn
-
-    async def init_db(self):
-        """Initialize the database schema if it doesn't exist.
-        This should already be initialized by the indexer, but we double-check here."""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            try:
-                cursor = await conn.cursor()
-
-                # Check if tables exist
-                await cursor.execute(
-                    "SELECT name FROM sqlite_master WHERE type='table' AND name='artists'"
-                )
-                if not await cursor.fetchone():
-                    logger.warning(
-                        "Database not initialized. Please run the indexer first."
-                    )
-
-                logger.info("Database check completed")
-            except Exception as e:
-                logger.error(f"Error initializing database: {str(e)}")
-                await conn.rollback()
-                raise
-
-    async def get_track_by_id(self, track_id: str) -> Optional[Dict]:
-        """Get track information by ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT t.*, a.title as album_title, ar.name as artist_name 
-                FROM tracks t
-                LEFT JOIN albums a ON t.album_id = a.id
-                LEFT JOIN artists ar ON t.artist_id = ar.id
-                WHERE t.id = ?
-            """,
-                (track_id,),
-            )
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_tracks_by_ids(self, track_ids: List[str]) -> List[Dict]:
-        """Get track information by IDs"""
-        if not track_ids:
-            return []
-
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            placeholders = ", ".join("?" for _ in track_ids)
-            await cursor.execute(
-                f"""
-                SELECT t.*, a.title as album_title, ar.name as artist_name 
-                FROM tracks t
-                LEFT JOIN albums a ON t.album_id = a.id
-                LEFT JOIN artists ar ON t.artist_id = ar.id
-                WHERE t.id IN ({placeholders})
-            """,
-                track_ids,
-            )
-            rows = await cursor.fetchall()
-            return [dict(row) for row in rows]
-
-    async def get_artist_by_id(self, artist_id: str) -> Optional[Dict]:
-        """Get artist information by ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute("SELECT * FROM artists WHERE id = ?", (artist_id,))
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_artist_by_mbid(self, mbid: str) -> Optional[Dict]:
-        """Get artist information by MusicBrainz ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                "SELECT * FROM artists WHERE mbid = ? LIMIT 1", (mbid,)
-            )
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_album_by_id(self, album_id: str) -> Optional[Dict]:
-        """Get album information by ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT a.*, ar.name as artist_name
-                FROM albums a
-                LEFT JOIN artists ar ON a.artist_id = ar.id
-                WHERE a.id = ?
-            """,
-                (album_id,),
-            )
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_album_by_mbid(self, mbid: str) -> Optional[Dict]:
-        """Get album information by MusicBrainz ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT a.*, ar.name as artist_name
-                FROM albums a
-                LEFT JOIN artists ar ON a.artist_id = ar.id
-                WHERE a.mbid = ? LIMIT 1
-                """,
-                (mbid,),
-            )
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_album_by_title_and_artist(
-        self, title: str, artist_id: str
-    ) -> Optional[Dict]:
-        """Get album information by title and artist ID"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT a.*, ar.name as artist_name
-                FROM albums a
-                LEFT JOIN artists ar ON a.artist_id = ar.id
-                WHERE a.title = ? AND a.artist_id = ? LIMIT 1
-                """,
-                (title, artist_id),
-            )
-            row = await cursor.fetchone()
-            return dict(row) if row else None
-
-    async def get_non_enriched_artists(self, limit: int = 50) -> List[Dict]:
-        """Get artists that haven't been enriched yet"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT * FROM artists
-                WHERE enriched = 0 AND id != 'unknown_artist'
-                LIMIT ?
-            """,
-                (limit,),
-            )
-            rows = await cursor.fetchall()
-            return [dict(row) for row in rows]
-
-    async def get_non_enriched_albums(self, limit: int = 50) -> List[Dict]:
-        """Get albums that haven't been enriched yet"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT a.*, ar.name as artist_name
-                FROM albums a
-                JOIN artists ar ON a.artist_id = ar.id
-                WHERE a.enriched = 0 AND a.id != 'unknown_album'
-                LIMIT ?
-            """,
-                (limit,),
-            )
-            rows = await cursor.fetchall()
-            return [dict(row) for row in rows]
-
-    async def get_non_enriched_tracks(self, limit: int = 50) -> List[Dict]:
-        """Get tracks that haven't been enriched yet"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            await cursor.execute(
-                """
-                SELECT t.*, a.title as album_title, ar.name as artist_name
-                FROM tracks t
-                LEFT JOIN albums a ON t.album_id = a.id
-                LEFT JOIN artists ar ON t.artist_id = ar.id
-                WHERE t.enriched = 0
-                LIMIT ?
-            """,
-                (limit,),
-            )
-            rows = await cursor.fetchall()
-            return [dict(row) for row in rows]
-
-    async def search_artists(
-        self, query: str, limit: int = 50
-    ) -> Tuple[List[Dict], int]:
-        """Search artists by query"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-            search_term = f"%{query}%"
-
-            # Get total count
-            await cursor.execute(
-                "SELECT COUNT(*) as count FROM artists WHERE name LIKE ?",
-                (search_term,),
-            )
-            row = await cursor.fetchone()
-            total = row["count"] if row else 0
-
-            # Get results
-            await cursor.execute(
-                """
-                SELECT * FROM artists
-                WHERE name LIKE ?
-                ORDER BY name
-                LIMIT ?
-            """,
-                (search_term, limit),
-            )
-            rows = await cursor.fetchall()
-            return [dict(row) for row in rows], total
-
-    async def update_artist(self, artist_id: str, data: Dict[str, Any]) -> None:
-        """Update artist information"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-
-            # Get artist table column names
-            await cursor.execute("PRAGMA table_info(artists)")
-            rows = await cursor.fetchall()
-            valid_columns = {row["name"] for row in rows}
-
-            # Filter out data keys that don't exist in the artists table
-            filtered_data = {k: v for k, v in data.items() if k in valid_columns}
-
-            if not filtered_data:
-                logger.debug(f"No valid columns to update for artist {artist_id}")
-                return
-
-            # Build the SET clause with only valid columns
-            fields = []
-            values = []
-            for key, value in filtered_data.items():
-                fields.append(f"{key} = ?")
-                values.append(value)
-
-            # Add artist_id to the values
-            values.append(artist_id)
-
-            query = f"UPDATE artists SET {', '.join(fields)} WHERE id = ?"
-            await cursor.execute(query, values)
-            await conn.commit()
-
-            # Log if any fields were filtered out
-            filtered_out = set(data.keys()) - valid_columns
-            if filtered_out:
-                logger.debug(
-                    f"Filtered out non-existent columns for artist {artist_id}: {', '.join(filtered_out)}"
-                )
-
-    async def update_album(self, album_id: str, data: Dict[str, Any]) -> None:
-        """Update album information"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-
-            # Get album table column names
-            await cursor.execute("PRAGMA table_info(albums)")
-            rows = await cursor.fetchall()
-            valid_columns = {row["name"] for row in rows}
-
-            # Filter out data keys that don't exist in the albums table
-            filtered_data = {k: v for k, v in data.items() if k in valid_columns}
-
-            if not filtered_data:
-                logger.debug(f"No valid columns to update for album {album_id}")
-                return
-
-            # Build the SET clause with only valid columns
-            fields = []
-            values = []
-            for key, value in filtered_data.items():
-                fields.append(f"{key} = ?")
-                values.append(value)
-
-            # Add album_id to the values
-            values.append(album_id)
-
-            query = f"UPDATE albums SET {', '.join(fields)} WHERE id = ?"
-            await cursor.execute(query, values)
-            await conn.commit()
-
-            # Log if any fields were filtered out
-            filtered_out = set(data.keys()) - valid_columns
-            if filtered_out:
-                logger.debug(
-                    f"Filtered out non-existent columns for album {album_id}: {', '.join(filtered_out)}"
-                )
-
-    async def update_track(self, track_id: str, data: Dict[str, Any]) -> None:
-        """Update track information"""
-        async with self._get_connection() as conn:
-            conn.row_factory = aiosqlite.Row
-            cursor = await conn.cursor()
-
-            # Get track table column names
-            await cursor.execute("PRAGMA table_info(tracks)")
-            rows = await cursor.fetchall()
-            valid_columns = {row["name"] for row in rows}
-
-            # Filter out data keys that don't exist in the tracks table
-            filtered_data = {k: v for k, v in data.items() if k in valid_columns}
-
-            if not filtered_data:
-                logger.debug(f"No valid columns to update for track {track_id}")
-                return
-
-            # Build the SET clause with only valid columns
-            fields = []
-            values = []
-            for key, value in filtered_data.items():
-                fields.append(f"{key} = ?")
-                values.append(value)
-
-            # Add track_id to the values
-            values.append(track_id)
-
-            query = f"UPDATE tracks SET {', '.join(fields)} WHERE id = ?"
-            await cursor.execute(query, values)
-            await conn.commit()
-
-            # Log if any fields were filtered out
-            filtered_out = set(data.keys()) - valid_columns
-            if filtered_out:
-                logger.debug(
-                    f"Filtered out non-existent columns for track {track_id}: {', '.join(filtered_out)}"
-                )
-
-    async def insert_artist(self, data: Dict[str, Any]) -> None:
-        """Insert a new artist"""
-        async with self._get_connection() as conn:
-            cursor = await conn.cursor()
-
-            # Build the query
-            fields = ", ".join(data.keys())
-            placeholders = ", ".join("?" for _ in data)
-            values = list(data.values())
-
-            query = f"INSERT OR REPLACE INTO artists ({fields}) VALUES ({placeholders})"
-            await cursor.execute(query, values)
-            await conn.commit()
-
-    async def insert_album(self, data: Dict[str, Any]) -> None:
-        """Insert a new album"""
-        async with self._get_connection() as conn:
-            cursor = await conn.cursor()
-
-            # Build the query
-            fields = ", ".join(data.keys())
-            placeholders = ", ".join("?" for _ in data)
-            values = list(data.values())
-
-            query = f"INSERT OR REPLACE INTO albums ({fields}) VALUES ({placeholders})"
-            await cursor.execute(query, values)
-            await conn.commit()
-
-    async def insert_track(self, data: Dict[str, Any]) -> None:
-        """Insert a new track"""
-        async with self._get_connection() as conn:
-            cursor = await conn.cursor()
-
-            # Build the query
-            fields = ", ".join(data.keys())
-            placeholders = ", ".join("?" for _ in data)
-            values = list(data.values())
-
-            query = f"INSERT OR REPLACE INTO tracks ({fields}) VALUES ({placeholders})"
-            await cursor.execute(query, values)
-            await conn.commit()
